+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Expose transaction interface on storage adapter [#285](https://github.com/rokwire/core-building-block/issues/285)
 
 ## [1.11.0] - 2021-11-04
 ### Added